\name{Bimap-methods}

\alias{Bimap-methods}

% Classes

\alias{Bimap}
\alias{class:Bimap}
\alias{Bimap-class}

% Methods

\alias{colnames}
\alias{colnames,FlatBimap-method}
\alias{colnames,AnnDbBimap-method}

\alias{colmetanames}
\alias{colmetanames,FlatBimap-method}
\alias{colmetanames,AnnDbBimap-method}


\title{Bimap methods}

\description{
  coming soon...
}

\usage{
  colnames(x, do.NULL=TRUE, prefix="col")
  colmetanames(x)

  ## many more Bimap methods coming soon...
}

\arguments{
  \item{x}{
    A Bimap object.
  }
  \item{do.NULL}{
    Ignored.
  }
  \item{prefix}{
    Ignored.
  }
}

\value{
}

\author{H. Pages}

\seealso{
}

\examples{
}

\keyword{methods}

