\name{AnnotationDbi-defunct}
\alias{AnnotationDbi-defunct}
\title{Defunct Functions in Package \pkg{AnnotationDbi}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{makeARABIDOPSISCHIP_DB}
\alias{makeCHICKENCHIP_DB}
\alias{makeCHICKENCHIP_DB}
\alias{makePIGCHIP_DB}
\alias{makeWORMCHIP_DB}
\alias{makeBOVINECHIP_DB}
\alias{makeCANINECHIP_DB}
\alias{makeECOLICHIP_DB}
\alias{makeZEBRAFISHCHIP_DB}
\alias{makeYEASTCHIP_DB}
\alias{makeFLYCHIP_DB}
\alias{makeRATCHIP_DB}
\alias{makeMOUSECHIP_DB}
\alias{makeHUMANCHIP_DB}
%
\description{
  The functions or variables listed here are no longer part of the AnnotationDbi package.
}
\seealso{
  \code{\link{Defunct}}
}


