\name{AnnotationDbi-deprecated}
\alias{AnnotationDbi-deprecated}
\title{Defunct Functions in Package \pkg{AnnotationDbi}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{makeRHESUSCHIP_DB}
%
\description{
  The functions or variables listed here have been deprecated and should no
  longer be used.
}
\seealso{
  \code{\link{Deprecated}}
}

