\name{AnnDbPkg-checker}

\alias{AnnDbPkg-checker}

\alias{checkMAPCOUNTS}


\title{Check the SQL data contained in an SQLite-based annotation package}

\description{
  Check the SQL data contained in an SQLite-based annotation package.
}

\usage{
  checkMAPCOUNTS(pkgname)
}

\arguments{
  \item{pkgname}{
    The name of the SQLite-based annotation package to check.
  }
}

\details{
}

\author{H. Pages}

\seealso{
  \link{AnnDbPkg-maker}
}

\examples{
  if (FALSE)
      checkMAPCOUNTS("hgu95av2.db")
}

\keyword{utilities}

