\name{AnnotationDbi internals}

\alias{debugSQL}

\alias{class:FlatBimap}
\alias{FlatBimap-class}
\alias{FlatBimap}

\alias{class:L2Rlink}
\alias{L2Rlink-class}
\alias{L2Rlink}

\alias{initialize,FlatBimap-method}
\alias{initialize,L2Rlink-method}
\alias{show,L2Rlink-method}

\alias{dbFileConnect}
\alias{dbFileDisconnect}
\alias{mergeToNamespaceAndExport}

\alias{createAnnObjs.HUMANCHIP_DB}
\alias{createAnnObjs.RODENTCHIP_DB}
\alias{createAnnObjs.FLYCHIP_DB}
\alias{createAnnObjs.YEASTCHIP_DB}
\alias{createAnnObjs.ARABIDOPSISCHIP_DB}
\alias{createAnnObjs.HUMAN_DB}
\alias{createAnnObjs.RODENT_DB}
\alias{createAnnObjs.FLY_DB}
\alias{createAnnObjs.YEAST_DB}
\alias{createAnnObjs.YEASTLEGACY_DB}
\alias{createAnnObjs.GO_DB}
\alias{createAnnObjs.KEGG_DB}
\alias{createAnnObjs.AFFYHUEX_DB}

\title{AnnotationDbi internals}

\description{
  AnnotationDbi objects, classes and methods that are not intended to
  be used directly.
}

\keyword{internal}
\keyword{methods}
\keyword{classes}
