
\name{AnnotationDbi-deprecated}
\alias{AnnotationDbi-deprecated}
\title{Deprecated Functions in package \pkg{AnnotationDbi}}

\alias{determineDefaultSeqnameStyle}
\alias{determineDefaultSeqnameStyle,TranscriptDb-method}
\alias{supportedSeqnameMappings}
\alias{listAllSupportedStylesBySpecies}
\alias{listAllSupportedSeqnameStyles}
\alias{supportedSeqnameStyles}
\alias{supportedSeqnames}
\alias{isSupportedSeqnamesStyle}
\alias{testSeqnames}
\alias{findSequenceRenamingMaps}
\alias{extractSeqnameSet}

\description{
  The functions or variables listed here have been deprecated and should no
  longer be used.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{supportedSeqnameMappings: \code{\link[GenomeInfoDb]{genomeStyles}}}
    \item{listAllSupportedStylesBySpecies: \code{\link[GenomeInfoDb]{genomeStyles}}}
    \item{listAllSupportedSeqnameStyles: \code{\link[GenomeInfoDb]{genomeStyles}}}
    \item{supportedSeqnameStyles: \code{\link[GenomeInfoDb]{genomeStyles}}}
    \item{supportedSeqnames: \code{\link[GenomeInfoDb]{genomeStyles}}}
    
    \item{isSupportedSeqnamesStyle}
    \item{testSeqnames}
    
    \item{findSequenceRenamingMaps: \code{\link[GenomeInfoDb]{mapSeqlevels}}}
    \item{extractSeqnameSet: \code{\link[GenomeInfoDb]{extractSeqlevels}}}
  }
}

\seealso{
  \code{\link{Deprecated}}
}
